+++ contracts/mocks/access/AccessControlMock.sol
@@ -20,6 +20,10 @@ contract AccessControlMock is AccessControl, ForwarderRegistryContext {
         AccessControlStorage.layout().enforceHasRole(role, account);
     }
 
+    function enforceHasTargetContractRole(address targetContract, bytes32 role, address account) external view {
+        AccessControlStorage.enforceHasTargetContractRole(targetContract, role, account);
+    }
+
     function __msgData() external view returns (bytes calldata) {
         return _msgData();
     }